% Auto-generated: do not edit by hand
\name{daqDarkThemeProvider}

\alias{daqDarkThemeProvider}

\title{DarkThemeProvider component}

\description{
DarkThemeProvider is a component that is placed at the root of the component tree to make all components match the dark theme
}

\usage{
daqDarkThemeProvider(children=NULL, theme=NULL)
}

\arguments{
\item{children}{The children of this component}

\item{theme}{Theme object to override with a custom theme}
}
