% Auto-generated: do not edit by hand
\name{daqIndicator}

\alias{daqIndicator}

\title{Indicator component}

\description{
A boolean indicator LED.
}

\usage{
daqIndicator(id=NULL, value=NULL, color=NULL, size=NULL, width=NULL, height=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify the indicator in Dash callbacks}

\item{value}{If true, indicator is illuminated}

\item{color}{Color of the indicator}

\item{size}{Size of the component. Either use this or width and height}

\item{width}{Width of the component. Set both width and height for a rectangular indicator}

\item{height}{Height of the component. Set both width and height for a rectangular indicator}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the control. To control styling,
pass an object with label and style properties}

\item{labelPosition}{Where the indicator label is positioned}

\item{className}{Class to apply to the root component element}

\item{style}{Style to apply to the root component element}
}
