% Auto-generated: do not edit by hand
\name{daqKnob}

\alias{daqKnob}

\title{Knob component}

\description{
A knob component that can be turned to a value between some range.
}

\usage{
daqKnob(id=NULL, value=NULL, color=NULL, size=NULL, min=NULL, max=NULL, disabled=NULL, theme=NULL, label=NULL, labelPosition=NULL, scale=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{value}{The value of knob}

\item{color}{Color configuration for the knob's track.}

\item{size}{The size (diameter) of the knob in pixels}

\item{min}{The minimum value of the knob}

\item{max}{The maximum value of the knob}

\item{disabled}{If true, knob cannot be moved.}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the control. To control styling, pass an object with label and style properties.}

\item{labelPosition}{Where the knob label is positioned.}

\item{scale}{Configuration for the component scale.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root component element.}
}
