% Auto-generated: do not edit by hand
\name{daqNumericInput}

\alias{daqNumericInput}

\title{NumericInput component}

\description{
A numeric input component that can be set to a value between some range.
}

\usage{
daqNumericInput(id=NULL, value=NULL, size=NULL, min=NULL, max=NULL, disabled=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{value}{The value of numeric input}

\item{size}{The size (length) of the numeric input in pixels}

\item{min}{The minimum value of the numeric input}

\item{max}{The maximum value of the numeric input}

\item{disabled}{If true, numeric input cannot changed.}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the control. To control styling,
pass an object with label and style properties.}

\item{labelPosition}{Where the numeric input label is positioned.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root component element.}
}
