% Auto-generated: do not edit by hand
\name{daqPrecisionInput}

\alias{daqPrecisionInput}

\title{PrecisionInput component}

\description{
A numeric input component that converts an input value to the desired precision.
}

\usage{
daqPrecisionInput(id=NULL, value=NULL, size=NULL, min=NULL, max=NULL, precision=NULL, disabled=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{value}{The value of numeric input}

\item{size}{The size (length) of the numeric input in pixels}

\item{min}{The minimum value of the numeric input}

\item{max}{The maximum value of the numeric input}

\item{precision}{Number of significant figures}

\item{disabled}{If true, numeric input cannot be changed.}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the scientific notation. To control styling,
pass an object with label and style properties.}

\item{labelPosition}{Where the numeric input label is positioned.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root component element.}
}
