% Auto-generated: do not edit by hand
\name{daqStopButton}

\alias{daqStopButton}

\title{StopButton component}

\description{
A Stop button component
}

\usage{
daqStopButton(children=NULL, id=NULL, size=NULL, buttonText=NULL, n_clicks=NULL, disabled=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{children}{The children of the button.}

\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{size}{The size (width) of the stop button in pixels}

\item{buttonText}{Text displayed in the button}

\item{n_clicks}{Number of times the button was clicked}

\item{disabled}{If true, button cannot be pressesd.}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the button.
To control styling, pass an object with label and
style properties.}

\item{labelPosition}{Where the  label is positioned.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root component element.}
}
