% Auto-generated: do not edit by hand
\docType{package}
\name{dashDaq-package}
\alias{dashDaq}
\title{Interactive Data Acquisition and Control Components for Dash }
\description{
A robust set of controls that make it simpler to integrate data acquisition and controls into your Dash applications.
}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plot.ly>
}
