% Auto-generated: do not edit by hand
\name{daqColorPicker}

\alias{daqColorPicker}

\title{ColorPicker component}

\description{
A color picker.
}

\usage{
daqColorPicker(id=NULL, value=NULL, disabled=NULL, size=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify the color picker in Dash callbacks}

\item{value}{Color value of the picker}

\item{disabled}{If true, color cannot be picked.}

\item{size}{Size (width) of the component in pixels}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the control. To control styling,
pass an object with label and style properties}

\item{labelPosition}{Where the indicator label is positioned}

\item{className}{Class to apply to the root component element}

\item{style}{Style to apply to the root component element}
}
