% Auto-generated: do not edit by hand
\name{daqGraduatedBar}

\alias{daqGraduatedBar}

\title{GraduatedBar component}

\description{
A graduated bar component that displays a value within some range as a percentage.
}

\usage{
daqGraduatedBar(id=NULL, value=NULL, color=NULL, size=NULL, vertical=NULL, min=NULL, max=NULL, step=NULL, showCurrentValue=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{value}{The value of the graduated bar}

\item{color}{Color configuration for the graduated bar's
progress blocks.}

\item{size}{The size (length) of the graduated bar in pixels}

\item{vertical}{If true, will display bar vertically instead of horizontally}

\item{min}{The minimum value of the graduated bar}

\item{max}{The maximum value of the graduated bar}

\item{step}{Value by which progress blocks appear}

\item{showCurrentValue}{If true, the current percentage
 of the bar will be displayed}

\item{label}{Description to be displayed alongside the control. To control styling, pass an object with label and style properties.}

\item{labelPosition}{Where the component label is positioned.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root component element.}
}
