% Auto-generated: do not edit by hand
\name{daqLEDDisplay}

\alias{daqLEDDisplay}

\title{LEDDisplay component}

\description{
A 7-bar LED display component.
}

\usage{
daqLEDDisplay(id=NULL, value=NULL, color=NULL, backgroundColor=NULL, size=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify the display in Dash callbacks}

\item{value}{Value to be displayed. A number or a string
containing only digits (0-9), periods, and colons.}

\item{color}{Color of the display}

\item{backgroundColor}{Color of the display's background}

\item{size}{Size of the display}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the control. To control styling,
pass an object with label and style properties}

\item{labelPosition}{Where the display label is positioned}

\item{className}{Class to apply to the root component element}

\item{style}{Style to apply to the root component element}
}
