% Auto-generated: do not edit by hand
\name{daqPowerButton}

\alias{daqPowerButton}

\title{PowerButton component}

\description{
A power button component can be turned on and off.
}

\usage{
daqPowerButton(id=NULL, on=NULL, color=NULL, size=NULL, disabled=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{on}{Whether or not the power button is on}

\item{color}{The indicator color to display when power button is on}

\item{size}{The size (diameter) of the power button in pixels}

\item{disabled}{If true, power button cannot be clicked}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the button. To control styling, pass an object with label and style properties.}

\item{labelPosition}{Where the button label is positioned.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root component element.}
}
