% Auto-generated: do not edit by hand
\name{daqToggleSwitch}

\alias{daqToggleSwitch}

\title{ToggleSwitch component}

\description{
A switch component that toggles between two values.
}

\usage{
daqToggleSwitch(id=NULL, value=NULL, size=NULL, color=NULL, vertical=NULL, disabled=NULL, theme=NULL, label=NULL, labelPosition=NULL, className=NULL, style=NULL)
}

\arguments{
\item{id}{The ID used to identify this compnent in Dash callbacks}

\item{value}{The state of the switch}

\item{size}{The size of the switch}

\item{color}{Color to highlight button/indicator}

\item{vertical}{If true, switch will be vertical instead
of horizontal}

\item{disabled}{If true, switch cannot be clicked}

\item{theme}{Theme configuration to be set by a ThemeProvider}

\item{label}{Description to be displayed alongside the control. To control styling, pass an object with label and style properties.}

\item{labelPosition}{Where the component label is positioned.}

\item{className}{Class to apply to the root component element.}

\item{style}{Style to apply to the root object.}
}
